added logo, changed styles, fixed web auth
