+++ cheaper-ui/package.json
@@ -10,6 +10,7 @@
     "@types/node": "^18.11.9",
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.8",
+    "@types/redux-logger": "^3.0.9",
     "antd": "^4.24.1",
     "eslint": "^8.27.0",
     "react": "^18.2.0",
@@ -17,6 +18,7 @@
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
+    "redux-logger": "^3.0.6",
     "sass": "^1.56.1",
     "typescript": "^4.8.4",
     "web-vitals": "^2.1.0"
