+++ cheaper-ui/src/base.sass
@@ -1,7 +1,6 @@
 
-.tc-header
-    a
-        margin: 0 10px
-
 .tc-page
     min-height: 1080px
+
+.tc-main
+    background: white
