+++ cheaper-ui/src/components/header/styles.sass
@@ -1,0 +1,6 @@
+
+.tc-header
+    background: white !important
+    box-shadow: 0 1px 2px 0 rgb(0 0 0 / 3%), 0 1px 6px -1px rgb(0 0 0 / 2%), 0 2px 4px 0 rgb(0 0 0 / 2%)
+    a
+        margin: 0 10px
