+++ cheaper-ui/src/components/login-page/login-form.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
-import { Button, Checkbox, Form, Input } from 'antd';
+import { Button, Checkbox, Col, Form, Input, Row, Typography } from 'antd';
 import { useDispatch } from 'react-redux';
 import { loginAsync } from '../../store/actions/auth';
 import { AppDispatch } from '../../store/store';
 import { LoginData } from '../../store/interfaces';
+
+const { Title } = Typography;
 
 export function LoginForm () {
     const dispatch: AppDispatch = useDispatch();
@@ -17,41 +19,50 @@ export function LoginForm () {
     };
 
     return (
-        <Form
-            name="basic"
-            labelCol={{ span: 8 }}
-            wrapperCol={{ span: 16 }}
-            initialValues={{ remember: true }}
-            onFinish={onFinish}
-            onFinishFailed={onFinishFailed}
-            autoComplete="off"
-            className='tc-login-form'
-        >
-        <Form.Item
-            label="Имя пользователя"
-            name="username"
-            rules={[{ required: true, message: 'Please input your username!' }]}
-        >
-            <Input />
-        </Form.Item>
+        <>
+            <Row justify='center'>
+                <Col>
+                    <Title level={2}>Вход</Title>
+                </Col>
+            </Row>
+            <Form
+                name="basic"
+                initialValues={{ remember: true }}
+                onFinish={onFinish}
+                onFinishFailed={onFinishFailed}
+                autoComplete="off"
+                className='tc-login-form'
+            >
+            <Form.Item
+                name="username"
+                rules={[{ required: true, message: 'Please input your username!' }]}
+            >
+                <Input 
+                    placeholder='Имя пользователя'
+                />
+            </Form.Item>
 
-        <Form.Item
-            label="Пароль"
-            name="password"
-            rules={[{ required: true, message: 'Please input your password!' }]}
-        >
-            <Input.Password />
-        </Form.Item>
+            <Form.Item
+                name="password"
+                rules={[{ required: true, message: 'Please input your password!' }]}
+            >
+                <Input.Password 
+                    placeholder='Пароль'            
+                />
+            </Form.Item>
 
-        <Form.Item name="remember" valuePropName="checked" wrapperCol={{ offset: 8, span: 16 }}>
-            <Checkbox>Запомнить меня</Checkbox>
-        </Form.Item>
+            <Form.Item name="remember" valuePropName="checked" wrapperCol={{ offset: 1, span: 16 }}>
+                <Checkbox>Запомнить меня</Checkbox>
+            </Form.Item>
 
-        <Form.Item wrapperCol={{ offset: 8, span: 16 }}>
-            <Button type="primary" htmlType="submit">
-                Вход
-            </Button>
-        </Form.Item>
-        </Form>
+            <Form.Item>
+                <Row justify='center'>
+                    <Button type="primary" htmlType="submit">
+                            Вход
+                    </Button>
+                </Row>
+            </Form.Item>
+            </Form>
+        </>
     );
 }
