+++ cheaper-ui/src/components/login-page/login-page.tsx
@@ -1,14 +1,10 @@
-import { Col, Row, Typography } from 'antd';
+import { Col, Row } from 'antd';
 import React, { useEffect } from 'react'
-import { loginAsync, getUserAsync} from '../../store/actions/auth';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { useNavigate } from "react-router-dom";
 import { LoginForm } from './login-form';
 import './styles.sass'
 import { AppState } from '../../store/interfaces';
-
-
-const { Title } = Typography;
 
 export function LoginPage() {
     const navigate = useNavigate();
@@ -18,14 +14,9 @@ export function LoginPage() {
     }, [user])
     return (
         <Row className='tc-page tc-login-page' justify='center'>
-            <Col span={7}>
+            <Col span={8}>
                 <Row justify='center'>
-                    <Col span={8}>
-                        <Title level={2}>Вход</Title>
-                    </Col>
-                </Row>
-                <Row justify='center'>
-                    <Col flex='auto'>
+                    <Col>
                         <LoginForm/>
                     </Col>
                 </Row>
