+++ cheaper-ui/src/components/register-page/styles.sass
@@ -1,2 +1,5 @@
 .tc-register-page
     margin-top: 180px
+
+.tc-register-title
+    display: inline-block
